+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/PParams.hs
@@ -97,6 +97,7 @@ import Cardano.Ledger.Plutus.CostModels (
   emptyCostModels,
   getCostModelLanguage,
   getCostModelParams,
+  mkCostModels,
  )
 import Cardano.Ledger.Plutus.ExUnits (
   ExUnits (..),
@@ -408,7 +409,7 @@ instance Ord OrdExUnits where
 -- | Parameters that were added in Alonzo
 data UpgradeAlonzoPParams f = UpgradeAlonzoPParams
   { uappCoinsPerUTxOWord :: !(HKD f CoinPerWord)
-  , uappCostModels :: !(HKD f CostModels)
+  , uappPlutusV1CostModel :: !(HKD f CostModel)
   , uappPrices :: !(HKD f Prices)
   , uappMaxTxExUnits :: !(HKD f ExUnits)
   , uappMaxBlockExUnits :: !(HKD f ExUnits)
@@ -442,7 +443,7 @@ instance Default (UpgradeAlonzoPParams StrictMaybe) where
   def =
     UpgradeAlonzoPParams
       { uappCoinsPerUTxOWord = SNothing
-      , uappCostModels = SNothing
+      , uappPlutusV1CostModel = SNothing
       , uappPrices = SNothing
       , uappMaxTxExUnits = SNothing
       , uappMaxBlockExUnits = SNothing
@@ -611,7 +612,7 @@ upgradeAlonzoPParams UpgradeAlonzoPParams {..} ShelleyPParams {..} =
     , appMinPoolCost = sppMinPoolCost
     , -- new in alonzo
       appCoinsPerUTxOWord = uappCoinsPerUTxOWord
-    , appCostModels = uappCostModels
+    , appCostModels = hkdMap (Proxy @f) (mkCostModels . Map.singleton PlutusV1) uappPlutusV1CostModel
     , appPrices = uappPrices
     , appMaxTxExUnits = hkdMap (Proxy @f) OrdExUnits uappMaxTxExUnits
     , appMaxBlockExUnits = hkdMap (Proxy @f) OrdExUnits uappMaxBlockExUnits